Build a Python GUI with Tkinter
Inherit from the Frame class to create the Application Window.

Create a method to add the Window frame title and add the Labels to the grid.

Use Entry Widgets to allow the user to enter a filename and X and Y Coordinates. Add Text Widgets for displaying statistics.

Add buttons and styling to allow users to graph and display statistics.

Create button event handlers and retrieve and set data using widget methods.